refactor: 배송원 수행 페이지 비대면 전달 컴포넌트 구조 변경
