+++ client/src/components/executeComponents/RemoteDelivery.tsx
@@ -1,6 +1,13 @@
 import React, { useState, useEffect, ChangeEvent, useRef  } from "react";
 import styled from "styled-components";
+import BottomConfirmBtn from "../bottomconfirmBtn";
 import { ExecutionComponentProps } from "../../pages/ExecutionPage";
+import { useExecutionState } from "../../pages/ExecutionPage";
+import WaitClientConfirm from "./WaitClientConfirm";
+import { WriteTransactionToBlockchain } from "../../utils/ExecuteOrderFromBlockchain";
+import { SendDataToAndroid } from "../../utils/SendDataToAndroid";
+import { useAccount } from "wagmi";
+import { checkIsDelivering } from "../../utils/ExecuteOrderFromBlockchain";
 
 const Div0 = styled.div`
     display: flex;
@@ -13,12 +20,10 @@ font-size: var(--font-md1);
 font-weight: bold;
 `;
 
-interface RemoteDelivery extends ExecutionComponentProps{
-  state : File | null | undefined,
-  setState : React.Dispatch<React.SetStateAction<File | null | undefined>>,
-}
-
-export default function RemoteDelivery({ state, setState, orderNum }: RemoteDelivery) {
+export default function RemoteDelivery({ orderNum }: ExecutionComponentProps) {
+  const { setShowComponent } = useExecutionState()
+  const { address } = useAccount()
+  const [file, setFile] = useState<File | null | undefined>(undefined);
     const fileInput = useRef<HTMLInputElement>(null);
     
     const handleDivClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
@@ -27,19 +32,64 @@ export default function RemoteDelivery({ state, setState, orderNum }: RemoteDeli
     
     const imageChange = () => {
       if (fileInput.current?.files !== null) {
-        setState(fileInput.current?.files[0])
+        setFile(fileInput.current?.files[0])
       }
     }
+
+    const postImage = async () => {
+      const formData = new FormData();
+
+      if (file !== null && file !== undefined && orderNum !== undefined) {
+          formData.append('uploadImage', file);
+          formData.append('orderNum', orderNum);
+          const response = await fetch(process.env.REACT_APP_SERVER_URL + "order-complete-image", {
+              method: 'POST',
+              body: formData,
+          })
+          const message = await response.json()
+          if (message.msg === "done") {
+              alert("전송완료")
+          }
+      } else {
+          alert("사진이 없습니다.")
+      }
+  }  
+
+    const deliveredRogic = async () => {
+      if (orderNum !== undefined) {
+          const wttb = new WriteTransactionToBlockchain(orderNum)
+          const sdta = new SendDataToAndroid(address)
+          try {
+              const result = await wttb.deliveredOrder()
+              console.log(result)
+              // 배송원 수행 중인 오더 확인 후 false값 전송
+              if (!(await checkIsDelivering(address))) {
+                  sdta.sendIsDelivering(false)
+              }
+              // 배송원 사진 업로드 로직 작성
+              await postImage()
+          } catch(e) {
+              console.log(e)
+          }
+          setShowComponent(<WaitClientConfirm />)
+      }
+  }
 
     return(
       <>
         <Div0 onClick={handleDivClick}>
           <input ref={fileInput} onChange={imageChange} type="file" name="uploadImage" accept="image/png, image/jpeg, image/jpg" capture="environment" style={{ display: 'none' }} />
-          {(state !== null && state !== undefined) ? (
+          {(file !== null && file !== undefined) ? (
             <div>
-              <img src={URL.createObjectURL(state)} alt="uploaded photo"  width="100%" />
+              <img src={URL.createObjectURL(file)} alt="uploaded photo"  width="100%" />
             </div>
           ) : <Sp0>위탁장소에 배송된 사진을 촬영해주세요.</Sp0>}
+          <BottomConfirmBtn
+                content="확인"
+                confirmLogic={ () => {
+                    deliveredRogic();
+                }} isDisabled={false}
+            />
         </Div0>
       </>
     )
